+++ libs/interpret/src/lib/MLIDashboard/Controls/MultiICEPlot/MultiICEPlot.styles.ts
@@ -39,17 +39,3 @@ export const multiIcePlotStyles: () => IProcessedStyleSet<IMultiIcePlotStyles> =
       }
     });
   };
-
-export const SpinButtonStyles = {
-  labelWrapper: { alignSelf: "center" },
-  root: {
-    display: "inline-flex",
-    float: "right",
-    selectors: {
-      "> div": {
-        maxWidth: "78px"
-      }
-    }
-  },
-  spinButtonWrapper: { maxWidth: "68px" }
-};
